Remove mutator file and update error strings
