+++ app/graphql/concerns/query_authorize_coach.rb
@@ -1,0 +1,12 @@
+module QueryAuthorizeCoach
+  include ActiveSupport::Concern
+
+  def query_authorized?
+    # Needs a valid faculty profile
+    return false if current_user&.faculty.blank?
+
+    return false if course&.school != current_school
+
+    current_user.faculty.courses.exists?(id: course)
+  end
+end
