+++ config/locales/en.yml
@@ -748,6 +748,15 @@ en:
       multiple_method_of_completion: More than one method of completion
       invalid_checklist_error: Checklist not valid
       checklist_items_exceeded_error: Checklist should have less than 15 items
+    create_grading:
+      success_notification: The submission has been marked as reviewed.
+      grade_recorded: Grades Recorded
+      submission_missing_error: "Unable to find submission with ID: %{submission_id}"
+      submission_graded_error: Submission already graded
+      evaluation_criteria_error: "Cannot grade Submission %{submission_id} without evaluation criteria"
+      invalid_checklist_shape_error: the shape of data in the submission checklist do not match the one sent with the review
+      invalid_checklist_values_error: The values for checklist items in the submission does not match with review data
+      invalid_grading_error: "Grading values supplied are invalid: %{grades_data}"
     update_course:
       success_notification: Course updated successfully!
   offline:
